+++ src/numerical/mod.rs
@@ -1,9 +1,9 @@
 //! Differential equations & Numerical Analysis tools
 
-pub mod bdf;
-pub mod gauss_legendre;
+//pub mod bdf;
+//pub mod gauss_legendre;
 pub mod interp;
-pub mod newton;
+//pub mod newton;
 pub mod ode;
 pub mod spline;
 pub mod utils;